refactor: move templates
